+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-27 11:16:07
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-07-29 10:46:45
+// Commit: refactor(core): extract reusable atomic widgets from core
 
 class coreComponentState {
-  final String commitRef = '3170';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-07-27T11:16:07+05:00');
+  final String commitRef = '3208';
+  final String lastMessage = 'refactor(core): extract reusable atomic widgets from core';
+  final DateTime timestamp = DateTime.parse('2026-07-29T10:46:45+05:00');
 }